requirements.txt
swarms
tickr-agent
pydantic
loguru
swarm-models
fastapi 
uvicorn 
requests
requests-oauthlib
python-dotenv
schedule

# Enhanced Trading Algorithm Dependencies
yfinance>=0.2.25
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Technical Analysis & Quantitative Finance
pandas-ta>=0.3.14b
quantlib>=1.31
arch>=6.2.0

# Machine Learning for Trading
lightgbm>=4.0.0
xgboost>=1.7.0
tensorflow>=2.13.0

# Async & Web APIs
aiohttp>=3.8.0
websockets>=11.0.0
httpx>=0.24.0

# Database & Storage
sqlalchemy>=2.0.0

# Performance & Monitoring
psutil>=5.9.0
memory-profiler>=0.61.0

# Additional Utilities
python-dateutil>=2.8.0
pytz>=2023.3
holidays>=0.29 